Change namespaceID type to string (#647)
#### Type of change

- API Change
 
#### Description

This PR changes the namespaceID type from int to string. While this
potentially sacrifices smaller transaction sizes, it simplifies the
namespace lifecycle as it removes the need to perform client-side
mapping between string-based and integer-based namespaceIDs.

#### Additional details (Optional)

<!--- Additional implementation details or comments to reviewers. -->
<!--- Summarize how the pull request was tested (if not obvious from
commit). -->

#### Related issues

https://github.ibm.com/decentralized-trust-research/fscx/issues/45

+++ sidecar/ledger/ledger_service_test.go
@@ -56,8 +56,8 @@ func TestLedgerService(t *testing.T) {
 	inputBlock <- blk1
 	inputBlock <- blk2
 
-	for i := range 3 {
+	for i := range uint64(3) {
 		blk := <-receivedBlocksFromLedgerService
-		require.Equal(t, uint64(i), blk.Header.Number)
+		require.Equal(t, i, blk.Header.Number)
 	}
 }
